+++ src/main/java/roomescape/auth/service/AuthService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.auth.dto.LoginMember;
 import roomescape.auth.dto.LoginRequest;
 import roomescape.auth.dto.LoginResponse;
-import roomescape.common.exception.EntityNotFoundException;
-import roomescape.common.exception.LoginFailException;
+import roomescape.common.exception.business.EntityNotFoundException;
+import roomescape.common.exception.business.LoginFailException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.MemberId;
 import roomescape.member.domain.Role;
