--- src/main/java/roomescape/common/exception/EntityNotFoundException.java
@@ -1,10 +1,0 @@
-package roomescape.common.exception;
-
-import java.util.NoSuchElementException;
-
-public class EntityNotFoundException extends NoSuchElementException {
-
-    public EntityNotFoundException(final String message) {
-        super(message);
-    }
-}
