--- src/main/java/roomescape/common/exception/PaymentBadRequestException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class PaymentBadRequestException extends RuntimeException{
-
-    public PaymentBadRequestException(final String message) {
-        super(message);
-    }
-}
