--- src/main/java/roomescape/common/exception/PaymentServerException.java
@@ -1,8 +1,0 @@
-package roomescape.common.exception;
-
-public class PaymentServerException extends RuntimeException{
-
-    public PaymentServerException(final String message) {
-        super(message);
-    }
-}
