+++ src/main/java/roomescape/common/exception/business/AlreadyInUseException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class AlreadyInUseException extends BusinessException {
+
+    public AlreadyInUseException(final String message) {
+        super(message);
+    }
+}
