+++ src/main/java/roomescape/common/exception/business/AuthenticationException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class AuthenticationException extends BusinessException {
+
+    public AuthenticationException(final String message) {
+        super(message);
+    }
+}
