+++ src/main/java/roomescape/common/exception/business/BusinessException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class BusinessException extends RuntimeException{
+
+    public BusinessException(final String message) {
+        super(message);
+    }
+}
