+++ src/main/java/roomescape/common/exception/business/EntityNotFoundException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class EntityNotFoundException extends BusinessException {
+
+    public EntityNotFoundException(final String message) {
+        super(message);
+    }
+}
