+++ src/main/java/roomescape/common/exception/business/LoginFailException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class LoginFailException extends BusinessException {
+
+    public LoginFailException(final String message) {
+        super(message);
+    }
+}
