+++ src/main/java/roomescape/common/exception/business/PaymentBadRequestException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class PaymentBadRequestException extends BusinessException {
+
+    public PaymentBadRequestException(final String message) {
+        super(message);
+    }
+}
