+++ src/main/java/roomescape/common/exception/business/PaymentServerException.java
@@ -1,0 +1,8 @@
+package roomescape.common.exception.business;
+
+public class PaymentServerException extends BusinessException {
+
+    public PaymentServerException(final String message) {
+        super(message);
+    }
+}
