+++ src/main/java/roomescape/member/service/MemberService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import jakarta.validation.Valid;
-import roomescape.common.exception.EntityNotFoundException;
+import roomescape.common.exception.business.EntityNotFoundException;
 import roomescape.member.domain.Email;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Password;
