+++ src/test/java/roomescape/auth/service/AuthServiceTest.java
@@ -13,8 +13,8 @@ import org.springframework.test.context.ActiveProfiles;
 import roomescape.auth.dto.LoginMember;
 import roomescape.auth.dto.LoginRequest;
 import roomescape.auth.dto.LoginResponse;
-import roomescape.common.exception.EntityNotFoundException;
-import roomescape.common.exception.LoginFailException;
+import roomescape.common.exception.business.EntityNotFoundException;
+import roomescape.common.exception.business.LoginFailException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.repository.MemberRepository;
